# PRMS-FIXED/docs/PRMS_OpenAPI_v1_0.yaml
openapi: 3.1.0
info:
  title: PRMS — Patient Referral Management System API
  version: 1.0.0
  description: |
    Production API for the Kenya Patient Referral Management System (PRMS).
    
    **Base URL:** `https://api.prms.health.go.ke/api/v1`
    
    **Authentication:** Bearer JWT (RS256). Obtain tokens via `/auth/login`.
    
    **Compliance:** Kenya Data Protection Act (KDPA) 2019.
  contact:
    name: PRMS Engineering Team
    email: engineering@prms.health.go.ke
  license:
    name: Proprietary
servers:
  - url: https://api.prms.health.go.ke/api/v1
    description: Production
  - url: https://staging-api.prms.health.go.ke/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 signed JWT. Expires in 15 minutes.

# ─────────────────────────────────────────────────────────────────────────────
# REUSABLE SCHEMAS
# ─────────────────────────────────────────────────────────────────────────────
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    MetaWithPagination:
      allOf:
        - $ref: '#/components/schemas/Meta'
        - type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
        meta:
          $ref: '#/components/schemas/Meta'

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    FacilityLevel:
      type: string
      enum: [Level 2, Level 3, Level 4, Level 5, Level 6]
...
    UserRole:
      type: string
      enum: [System Admin, Hospital Admin, Clinician, Receptionist]
    UserStatus:
      type: string
      enum: [Active, Inactive, Suspended]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        hospitalId:
          type: integer
        status:
          $ref: '#/components/schemas/UserStatus'
        isTwoFactorEnabled:
          type: boolean
        isFirstLogin:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── Patient ───────────────────────────────────────────────────────────────
    PatientMasked:
      type: object
      description: Patient record with PII masked (Receptionist view)
      properties:
        id:
          type: integer
        fullName:
          type: string
          example: Jane W. M.
        nationalId:
          type: string
          example: XXXX6789
        gender:
          type: string
          enum: [Male, Female, Other]
        age:
          type: integer
        county:
          type: string
        phoneNumber:
          type: string
          example: 071XXXX678
...
    ReferralStatus:
      type: string
      enum: [Draft, Dispatched, Received, Accepted, Rejected, Completed]
    UrgencyLevel:
      type: string
      enum: [Routine, Urgent, Emergent]
...
    ReferralSummary:
      type: object
      properties:
        id:
          type: integer
        referralCode:
          type: string
          example: REF-2026-00981
        status:
          $ref: '#/components/schemas/ReferralStatus'
        urgencyLevel:
          $ref: '#/components/schemas/UrgencyLevel'
        direction:
          type: string
          enum: [incoming, outgoing]
        patient:
          $ref: '#/components/schemas/PatientMasked'
        sourceHospital:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            facilityLevel:
              $ref: '#/components/schemas/FacilityLevel'
        destinationHospital:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            facilityLevel:
              $ref: '#/components/schemas/FacilityLevel'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_TOKEN_EXPIRED
              message: Access token has expired
            meta:
              timestamp: "2026-06-10T10:00:00.000Z"
              requestId: uuid
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request body validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /auth/change-password:
    patch:
      tags: [Authentication]
      summary: Change own password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword, confirmPassword]
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 12
                confirmPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Password changed
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users:
    get:
      tags: [Users]
      summary: List facility staff (Hospital Admin)
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
        - name: role
          in: query
          schema:
            type: string
            enum: [Clinician, Receptionist]
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  meta:
                    $ref: '#/components/schemas/MetaWithPagination'

    post:
      tags: [Users]
      summary: Create a staff account (Hospital Admin)
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, username, email, phoneNumber, role]
              properties:
                fullName:
                  type: string
                username:
                  type: string
                email:
                  type: string
                  format: email
                phoneNumber:
                  type: string
                role:
                  type: string
                  enum: [Clinician, Receptionist]
      responses:
        '201':
          description: User created
        '409':
          description: Username or email already taken
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user details
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Users]
      summary: Update user profile
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                phoneNumber:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: User updated
...
  /referrals:
    post:
      tags: [Referrals]
      summary: Create a referral (Draft)
      operationId: createReferral
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientId, destinationHospitalId, urgencyLevel, reasonForReferral, clinicalSummary]
              properties:
                patientId:
                  type: integer
                destinationHospitalId:
                  type: integer
                urgencyLevel:
                  $ref: '#/components/schemas/UrgencyLevel'
                reasonForReferral:
                  type: string
                  minLength: 20
                clinicalSummary:
                  type: string
                  minLength: 50
      responses:
        '201':
          description: Referral created as Draft
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ReferralSummary'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Business rule violation

    get:
      tags: [Referrals]
      summary: List referrals (scoped to facility)
      operationId: listReferrals
      security:
        - BearerAuth: []
      parameters:
        - name: direction
          in: query
          schema:
            type: string
            enum: [incoming, outgoing]
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReferralStatus'
        - name: urgencyLevel
          in: query
          schema:
            $ref: '#/components/schemas/UrgencyLevel'
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Referral list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReferralSummary'
                  meta:
                    $ref: '#/components/schemas/MetaWithPagination'
...